// Admin/src/models/Threshold.Model.ts
export interface IThresholdModel {

  Active: any;
  AnswerType: any;
  AppreciateCalculation: any;
  CategoryId: any;
  Created: any;
  CreatedBy: any;
  DepreciateCalculation: any;
  DisableWarranty: any
  DisplayInList: any;
  DisplayName: any;
  Enabled: any;
  EnumName: any;
  Icons: any;
  Id: any;
  Identifier: any;
  IsEditable: any;
  IsValid: any;
  ModelVariantId: any;
  Modified: any;
  ModifiedBy: any;
  Name: any;
  NextQuestionNo: any;
  NextQuestionYes: any;
  OSTypeId: any;
  ParentId: any;
  ProductTypeId: any;
  QuestionnaireTypeId: any;
  QuestionnaireTypeName: any;
  Required: true;
  ResponseNo: any;
  ResponseYes: any;
  RowOrder: any;
  SubHeading: any;
  Threshold: any;
  ThresholdLevel: any;
  ThumbnailPath: any;
  Type: any;
  ValidationErrors: { Items: [] };
}